requirements.txt
argon2-cffi==20.1.0
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
commonmark==0.9.1
cycler==0.10.0
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
entrypoints==0.3
filelock==3.0.12
huggingface-hub==0.0.12
idna==3.2
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keybert==0.4.0
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.2
nltk==3.6.2
notebook==6.4.0
numpy==1.21.1
packaging==21.0
pandas==1.3.1
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.1
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
pycparser==2.20
Pygments==2.9.0
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
pywin32==301
pywinpty==1.1.3
PyYAML==5.4.1
pyzmq==22.1.0
qtconsole==5.1.1
QtPy==1.9.0
regex==2021.7.6
requests==2.26.0
rich==10.6.0
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.7.0
Send2Trash==1.7.1
sentence-transformers==2.0.0
sentencepiece==0.1.96
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
stanfordnlp==0.2.0
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.2.0
tokenizers==0.10.3
torch==1.9.0
torchvision==0.10.0
tornado==6.1
tqdm==4.61.2
traitlets==5.0.5
transformers==4.9.1
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
